backend/requirements.txt.txt
Flask
flask-cors
python-dotenv
openai
newspaper3k
gunicorn
